refactor: use the `MapKey` symbol from `utils` 🥳
