+++ src/controls/VControlScale.vue
@@ -2,8 +2,7 @@
   import { ScaleControl } from 'mapbox-gl';
   import type { PropType } from 'vue';
   import { defineComponent, onMounted } from 'vue';
-  import { MapKey } from '../../types/symbols';
-  import { injectStrict } from '../utils';
+  import { injectStrict, MapKey } from '../utils';
 
   export default defineComponent({
     name: 'VControlFullscreen',
