+++ src/utils/symbols.ts
@@ -1,0 +1,6 @@
+import type { Map } from 'mapbox-gl';
+import type { InjectionKey, Ref } from 'vue';
+
+const MapKey: InjectionKey<Ref<Map>> = Symbol('Map');
+
+export { MapKey };
